# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Game World Hub API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /parties:
    get:
      operationId: listParties
      tags: [parties]
      summary: List public parties and parties you are a member of
      responses:
        "200":
          description: Parties
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Party"
    post:
      operationId: createParty
      tags: [parties]
      summary: Create a party
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PartyInput"
      responses:
        "201":
          description: Party created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Party"
  /parties/{partyId}:
    get:
      operationId: getParty
      tags: [parties]
      summary: Get party details
      parameters:
        - name: partyId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Party
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Party"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateParty
      tags: [parties]
      summary: Update party (leader only)
      parameters:
        - name: partyId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PartyUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Party"
    delete:
      operationId: deleteParty
      tags: [parties]
      summary: Delete party (leader only)
      parameters:
        - name: partyId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /parties/{partyId}/invite:
    post:
      operationId: inviteToParty
      tags: [parties]
      summary: Invite a friend to the party
      parameters:
        - name: partyId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PartyInviteInput"
      responses:
        "200":
          description: Invited
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /parties/{partyId}/join:
    post:
      operationId: joinParty
      tags: [parties]
      summary: Join a party
      parameters:
        - name: partyId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Joined
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Party"
  /parties/{partyId}/leave:
    post:
      operationId: leaveParty
      tags: [parties]
      summary: Leave a party
      parameters:
        - name: partyId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Left
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /party-invites/{inviteId}/accept:
    post:
      operationId: acceptPartyInvite
      tags: [parties]
      summary: Accept a party invite
      parameters:
        - name: inviteId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Party"
  /party-invites/{inviteId}/decline:
    post:
      operationId: declinePartyInvite
      tags: [parties]
      summary: Decline a party invite
      parameters:
        - name: inviteId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Declined
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ── Games ─────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success

    # ── LFG (Looking For Group) ───────────────────────────────────────────────
...
    User:
      type: object
      required: [id, username, displayName, status, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        displayName:
          type: string
        avatarUrl:
          type: ["string", "null"]
        bannerUrl:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        emailVerified:
          type: boolean
        twoFactorMethod:
          type: string
          enum: [none, email, totp]
        allowProfileComments:
          type: boolean
        status:
          type: string
          enum: [online, away, busy, offline]
        currentGame:
          type: ["string", "null"]
        createdAt:
          type: string
...
    Party:
      type: object
      required: [id, name, leader, members, maxSize, isPublic, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        game:
          type: ["string", "null"]
        platform:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        leader:
          $ref: "#/components/schemas/User"
        members:
          type: array
          items:
            $ref: "#/components/schemas/User"
        maxSize:
          type: integer
        isPublic:
          type: boolean
        conversationId:
          type: ["integer", "null"]
        createdAt:
          type: string
    PartyInput:
      type: object
      required: [name, maxSize]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        game:
          type: string
        platform:
          type: string
        description:
          type: string
          maxLength: 500
        maxSize:
          type: integer
          minimum: 2
          maximum: 100
        isPublic:
          type: boolean
    PartyUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        game:
          type: string
        platform:
          type: string
        description:
          type: string
          maxLength: 500
        maxSize:
          type: integer
          minimum: 2
          maximum: 100
        isPublic:
          type: boolean
    PartyInviteInput:
      type: object
      required: [userId]
      properties:
        userId:
          type: integer
...